finished splitting out objectbox stuff.  also split out the main ReCall function so there's not an app.dart
