+++ lib/main.dart
@@ -1,85 +1,22 @@
+// lib/main.dart
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:recall/blocs/contact_list/contact_list_bloc.dart';
 import 'package:recall/repositories/contact_repository.dart';
-import 'package:recall/screens/contact_list_screen.dart';
-import 'package:recall/blocs/contact_details/contact_details_bloc.dart';
-import 'package:recall/screens/contact_details_screen.dart';
 import 'package:logger/logger.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:flutter/foundation.dart' show kIsWeb;
-import 'objectbox.g.dart'; // Import the generated ObjectBox model
+import 'package:recall/utils/objectbox_utils.dart' as objectbox_utils;
+import 'package:recall/app.dart';
+import 'objectbox.g.dart' as objectbox_g;
 
 final logger = Logger();
 
-late final Store? store; // Declare the store
-
-Future<Store?> openStore() async {
-  if (kIsWeb) {
-    return null;
-  }
-  try {
-    final docsDir = await getApplicationDocumentsDirectory();
-    return Store(getObjectBoxModel(), directory: '${docsDir.path}/objectbox');
-  } catch (e) {
-    logger.i("Error opening objectbox: $e");
-    return null;
-  }
-}
+late final objectbox_g.Store? store; // Declare the store
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized(); // Ensure Flutter is initialized
 
-  store = await openStore(); // Initialize the store
+  store = await objectbox_utils.openStore(); // Initialize the store
 
   runApp(ReCall(
       contactRepository:
           ContactRepository(store))); // Pass the store to the repository
   logger.i('LOG:App started');
-}
-
-class ReCall extends StatelessWidget {
-  final ContactRepository _contactRepository; // Add contactRepository parameter
-
-  const ReCall({super.key, required ContactRepository contactRepository})
-      : _contactRepository = contactRepository;
-  @override
-  Widget build(BuildContext context) {
-    return RepositoryProvider.value(
-      // Use RepositoryProvider.value
-      value: _contactRepository, // Provide the initialized repository
-      child: MultiBlocProvider(
-        providers: [
-          BlocProvider(
-            create: (context) => ContactListBloc(
-              contactRepository:
-                  _contactRepository, // Access repository directly
-            )..add(const ContactListEvent.loadContacts()),
-          ),
-          BlocProvider(
-            create: (context) => ContactDetailsBloc(
-              contactRepository:
-                  _contactRepository, // Access repository directly
-            ),
-          ),
-        ],
-        child: MaterialApp(
-          // Set the title of the app.
-          title: 'Contact App',
-          // Set the theme of the app.
-          theme: ThemeData(
-            primarySwatch: Colors.blue,
-          ),
-          // Set the initial route of the app.
-          home: const ContactListScreen(),
-          // Define the routes for the app.
-          routes: {
-            '/contactDetails': (context) =>
-                const ContactDetailsScreen(contactId: 0),
-          },
-          // ... (rest of your MaterialApp) ...
-        ),
-      ),
-    );
-  }
 }
